Revert "refactor: Rename consensus tasks and split storage (BFT-476) (#2357)"
This reverts commit 107e1a722d3e9812d96d829072346c5400af977b.
