+++ core/bin/external_node/src/main.rs
@@ -286,7 +286,7 @@ async fn run_core(
             // but we only need to wait for stop signal once, and it will be propagated to all child contexts.
             let ctx = ctx::root();
             scope::run!(&ctx, |ctx, s| async move {
-                s.spawn_bg(consensus::era::run_external_node(
+                s.spawn_bg(consensus::era::run_en(
                     ctx,
                     cfg,
                     pool,
